%f and %e are totaly finished. need to perfectly understand how %g works

+++ srcs/convert/integer.c
@@ -6,7 +6,7 @@
 /*   By: abaurens <abaurens@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/12/07 18:20:02 by abaurens          #+#    #+#             */
-/*   Updated: 2018/12/31 12:56:18 by abaurens         ###   ########.fr       */
+/*   Updated: 2019/01/04 14:33:18 by abaurens         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -30,10 +30,10 @@ static char		*integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (int)arg->value >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char		*long_integer(t_printf *const data, t_arg *const arg)
@@ -51,10 +51,10 @@ static char		*long_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (long int)arg->value >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char		*long_long_integer(t_printf *const data, t_arg *const arg)
@@ -72,10 +72,10 @@ static char		*long_long_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (long long)arg->value >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char		*short_integer(t_printf *const data, t_arg *const arg)
@@ -93,10 +93,10 @@ static char		*short_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (short int)arg->value >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char		*short_short_integer(t_printf *const data, t_arg *const arg)
@@ -114,10 +114,10 @@ static char		*short_short_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (signed char)arg->value >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char		*intmax_integer(t_printf *const data, t_arg *const arg)
@@ -135,10 +135,10 @@ static char		*intmax_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (intmax_t)arg->value >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char		*size_integer(t_printf *const data, t_arg *const arg)
@@ -156,10 +156,10 @@ static char		*size_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (long long)v >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char		*ssize_integer(t_printf *const data, t_arg *const arg)
@@ -177,10 +177,10 @@ static char		*ssize_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (long long)v >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char		*ptrdiff_integer(t_printf *const data, t_arg *const arg)
@@ -198,10 +198,10 @@ static char		*ptrdiff_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & (F_PLUS | F_SPAC)) && (ptrdiff_t)arg->value >= 0)
 		tab[ft_idxof('0', tab)] = (arg->flags & F_PLUS) ? '+' : ' ';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static const t_converter	g_funcs[] =
