+++ srcs/convert/strerror.c
@@ -33,7 +33,7 @@ static char		*character_string(t_printf *data, t_arg *arg)
 	ft_memset(res, (arg->flags & F_ZERO) ? '0' : ' ', tab_len);
 	tab_len -= ((arg->flags & F_MINS) ? tab_len : len);
 	ft_strncpy(res + tab_len, v, len);
-	res = (char *)ft_freturn(res, (long long)ft_strmcat(data->buf, res, -1));
+	res = (char *)ft_freturn(res, (long)ft_strmcat(data->buf, res, -1));
 	if (!res)
 		return (NULL);
 	free(data->buf);
