+++ srcs/convert/u_integer.c
@@ -26,10 +26,10 @@ static char			*integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*long_integer(t_printf *const data, t_arg *const arg)
@@ -43,10 +43,10 @@ static char			*long_integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*long_long_integer(t_printf *const data, t_arg *const arg)
@@ -60,10 +60,10 @@ static char			*long_long_integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*short_integer(t_printf *const data, t_arg *const arg)
@@ -77,10 +77,10 @@ static char			*short_integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*short_short_integer(t_printf *const data, t_arg *const arg)
@@ -94,10 +94,10 @@ static char			*short_short_integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*intmax_integer(t_printf *const data, t_arg *const arg)
@@ -111,10 +111,10 @@ static char			*intmax_integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*size_integer(t_printf *const data, t_arg *const arg)
@@ -128,10 +128,10 @@ static char			*size_integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*ssize_integer(t_printf *const data, t_arg *const arg)
@@ -145,10 +145,10 @@ static char			*ssize_integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*ptrdiff_integer(t_printf *const data, t_arg *const arg)
@@ -162,10 +162,10 @@ static char			*ptrdiff_integer(t_printf *const data, t_arg *const arg)
 	if (!(tab = padded_ulltoa(v, arg->precision, arg->min_width,
 		(arg->flags & F_MINS) != 0)))
 		return (NULL);
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static const t_converter	g_funcs[] =
