+++ srcs/convert/u_integer_hexahup.c
@@ -33,10 +33,10 @@ static char			*integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*long_integer(t_printf *const data, t_arg *const arg)
@@ -57,10 +57,10 @@ static char			*long_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*long_long_integer(t_printf *const data, t_arg *const arg)
@@ -81,10 +81,10 @@ static char			*long_long_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*short_integer(t_printf *const data, t_arg *const arg)
@@ -105,10 +105,10 @@ static char			*short_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*short_short_integer(t_printf *const data, t_arg *const arg)
@@ -129,10 +129,10 @@ static char			*short_short_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*intmax_integer(t_printf *const data, t_arg *const arg)
@@ -153,10 +153,10 @@ static char			*intmax_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*size_integer(t_printf *const data, t_arg *const arg)
@@ -177,10 +177,10 @@ static char			*size_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*ssize_integer(t_printf *const data, t_arg *const arg)
@@ -201,10 +201,10 @@ static char			*ssize_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static char			*ptrdiff_integer(t_printf *const data, t_arg *const arg)
@@ -225,10 +225,10 @@ static char			*ptrdiff_integer(t_printf *const data, t_arg *const arg)
 		return (NULL);
 	if ((arg->flags & F_HASH) && v != 0)
 		tab[ft_idxof('0', tab) + 1] = 'X';
-	tab = (char *)ft_freturn(tab, (long long)ft_strmcat(data->buf, tab, -1));
+	tab = (char *)ft_freturn(tab, (long)ft_strmcat(data->buf, tab, -1));
 	if (!tab)
 		return (NULL);
-	return (data->buf = (char *)ft_freturn(data->buf, (long long)tab));
+	return (data->buf = (char *)ft_freturn(data->buf, (long)tab));
 }
 
 static const t_converter	g_funcs[] =
